+++ src/main/java/org/geeksforgeeks/digitallibrary/repository/jpa/UserJPARepository.java
@@ -3,5 +3,5 @@ package org.geeksforgeeks.digitallibrary.repository.jpa;
 import org.geeksforgeeks.digitallibrary.entity.UserEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface UserJPARepository extends JpaRepository<UserEntity, Integer> {
+public interface UserJPARepository extends JpaRepository<UserEntity, Long> {
 }
